Fix Bug for Jumping to Calculator Mode Prematurely
Before this fix, if you have a parameter for a script that has both
a math symbol and number in it Mutate would jump to calculator mode
and no longer stick with the script it was previously matching. This
was because the calculator check happens before script searches, and
once the script check happened the keyword would be altered.

This prevents calculator checks from happening if the the buffer
matches a script.  This was compiled and tested on Ubuntu 15.04.